gnrc_sixlowpan_frag: move reassembly buffer to its own header
